% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{df_rect_board_tiles}
\alias{df_rect_board_tiles}
\title{Compute rectangular game boards using piecepack tiles}
\usage{
df_rect_board_tiles(
  nrows = 8,
  ncols = 8,
  x0 = 1,
  y0 = 1,
  max_tiles = 24,
  suit = rep(1:4, 6),
  rank = rep(1:6, each = 4)
)
}
\arguments{
\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{x0}{X coordinate for the center of the first cell/point}

\item{y0}{Y coordinate for the center of the first cell/point}

\item{max_tiles}{Maximum number of tiles that can be used}

\item{suit}{Vector of suit values to use for tile back (will be repeated).}

\item{rank}{Vector of rank values to use for tile back (will be repeated).}
}
\description{
\code{df_rect_board_tiles} returns a tibble \code{data_frame} of a rectangular board of desired size
made using a maximum number of piecepack tiles.
It will use either \dQuote{cells}, \dQuote{points}, and/or \dQuote{rivers} as necessary.
}
