% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R, R/game_starts_chess.R,
%   R/game_starts_traditional.R
\name{df_game}
\alias{df_game}
\alias{df_alien_city}
\alias{df_black_pawn_trucking}
\alias{df_brain_burn}
\alias{df_burbuja}
\alias{df_cardinals_guards}
\alias{df_cell_management}
\alias{df_change_change}
\alias{df_chariots}
\alias{df_chinese_checkers}
\alias{df_piecepack_halma}
\alias{df_mini_halma}
\alias{df_coin_collectors}
\alias{df_climbing_man}
\alias{df_crocodile_hop}
\alias{df_desfases}
\alias{df_easy_slider}
\alias{df_evade}
\alias{df_everest}
\alias{df_four_blind_mice}
\alias{df_froggy_bottom}
\alias{df_fujisan}
\alias{df_galaxy_express}
\alias{df_iceberg}
\alias{df_ice_floe}
\alias{df_japan}
\alias{df_lab_rats}
\alias{df_landlocked}
\alias{df_ley_lines}
\alias{df_mathrix}
\alias{df_piecepackman}
\alias{df_one_man_thrag}
\alias{df_pass_the_food}
\alias{df_piecepack_klondike}
\alias{df_piece_gaps}
\alias{df_piece_packing_pirates}
\alias{df_plans_of_action}
\alias{df_quatri}
\alias{df_relativity}
\alias{df_san_andreas}
\alias{df_sarcophagus}
\alias{df_shopping_mall}
\alias{df_skyscrapers}
\alias{df_slides_of_action}
\alias{df_speedy_towers}
\alias{df_steppin_stones}
\alias{df_the_in_crowd}
\alias{df_the_magic_bag}
\alias{df_the_penguin_game}
\alias{df_tower_of_babel}
\alias{df_tracers}
\alias{df_triactor}
\alias{df_tula}
\alias{df_wormholes}
\alias{df_alice_chess}
\alias{df_chaturaji}
\alias{df_four_seasons_chess}
\alias{df_international_chess}
\alias{df_chess}
\alias{df_minishogi}
\alias{df_shogi}
\alias{df_ultima}
\alias{df_baroque_chess}
\alias{df_xiangqi}
\alias{df_alquerque}
\alias{df_american_checkers}
\alias{df_awithlaknannai_mosona}
\alias{df_english_draughts}
\alias{df_checkers}
\alias{df_backgammon}
\alias{df_brandubh}
\alias{df_breakthrough}
\alias{df_cribbage_board}
\alias{df_cribbage}
\alias{textGrob_cribbage_board}
\alias{df_crossings}
\alias{df_four_field_kono}
\alias{df_grasshopper}
\alias{df_jul_gono}
\alias{df_lines_of_action}
\alias{df_ludo}
\alias{df_nine_mens_morris}
\alias{df_turkish_draughts}
\alias{df_twelve_mens_morris}
\alias{df_salta}
\alias{df_tablut}
\title{Data frames of starting diagrams for various games}
\usage{
df_alien_city(seed = NULL, tiles = NULL)

df_black_pawn_trucking(seed = NULL)

df_brain_burn(seed = NULL)

df_burbuja(seed = NULL, tiles = NULL)

df_cardinals_guards(seed = NULL, tiles = NULL)

df_cell_management(seed = NULL)

df_change_change(seed = NULL, coins = NULL)

df_chariots()

df_chinese_checkers()

df_piecepack_halma()

df_mini_halma()

df_coin_collectors(seed = NULL, tiles = NULL, coins = NULL, dice = NULL)

df_climbing_man(seed = NULL, variant = c("Basic", "Free"))

df_crocodile_hop(seed = NULL)

df_desfases(seed = NULL, tiles = NULL, dice = NULL)

df_easy_slider(seed = NULL, tiles = NULL, coins = NULL, pawns = NULL)

df_evade()

df_everest()

df_four_blind_mice()

df_froggy_bottom()

df_fujisan(seed = NULL, coins = NULL, dice = NULL)

df_galaxy_express(seed = NULL)

df_iceberg(seed = NULL, tiles = NULL)

df_ice_floe()

df_japan(seed = NULL)

df_lab_rats(seed = NULL)

df_landlocked(seed = NULL)

df_ley_lines()

df_mathrix(seed = NULL, coins = NULL)

df_piecepackman(seed = NULL, variant = "Roundabout")

df_one_man_thrag(seed = NULL)

df_pass_the_food()

df_piecepack_klondike(seed = NULL)

df_piece_gaps(seed = NULL)

df_piece_packing_pirates(seed = NULL)

df_plans_of_action(seed = NULL, coins = NULL)

df_quatri()

df_relativity(seed = NULL, coins = NULL)

df_san_andreas()

df_sarcophagus(seed = NULL)

df_shopping_mall(seed = NULL, cfg2 = "go")

df_skyscrapers(seed = NULL, tiles = NULL)

df_slides_of_action()

df_speedy_towers(n_players = 2, seed = NULL)

df_steppin_stones(seed = NULL)

df_the_in_crowd()

df_the_magic_bag(seed = NULL)

df_the_penguin_game(seed = NULL)

df_tower_of_babel(seed = NULL, tiles = NULL)

df_tracers()

df_triactor(seed = NULL, cfg2 = "playing_cards_expansion")

df_tula(
  seed = NULL,
  tiles = NULL,
  variant = c("Original", "Variant 1", "Variant 2", "Variant 3", "Variant 4")
)

df_wormholes()

df_alice_chess(has_subpack = FALSE, max_tiles = 24)

df_chaturaji(has_subpack = FALSE)

df_four_seasons_chess(has_subpack = FALSE)

df_international_chess(has_subpack = FALSE)

df_chess(has_subpack = FALSE)

df_minishogi()

df_shogi(has_subpack = FALSE, cfg2 = "piecepack")

df_ultima(has_subpack = FALSE)

df_baroque_chess(has_subpack = FALSE)

df_xiangqi(has_subpack = FALSE)

df_alquerque(has_matchsticks = FALSE)

df_american_checkers()

df_awithlaknannai_mosona(has_matchsticks = FALSE)

df_english_draughts()

df_checkers()

df_backgammon()

df_brandubh()

df_breakthrough()

df_cribbage_board()

df_cribbage()

textGrob_cribbage_board(suit_colors = rep("black", 4))

df_crossings()

df_four_field_kono()

df_grasshopper()

df_jul_gono()

df_lines_of_action()

df_ludo()

df_nine_mens_morris(has_matchsticks = FALSE)

df_turkish_draughts()

df_twelve_mens_morris(has_matchsticks = FALSE)

df_salta(has_subpack = FALSE)

df_tablut(die_width = 0.63)
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{tiles}{String of tile layout}

\item{coins}{String of coin layout}

\item{dice}{String of dice layout}

\item{variant}{Name of variant}

\item{pawns}{String of pawns layout}

\item{cfg2}{A string of a piecepack expansion (or perhaps \code{"piecepack"} for a second piecepack)}

\item{n_players}{Number of players}

\item{has_subpack}{Has a piecepack subpack}

\item{max_tiles}{Maximum number of (piecepack) tiles available to build boards}

\item{has_matchsticks}{Has matchsticks}

\item{suit_colors}{Character vector of the suit colors}

\item{die_width}{Width of dice}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
data frame output can usually be plotted with \code{pmap_piece(df, cfg = cfg, default.units = "in")}.
}
\details{
Here is some more information about the various games:
\describe{
\item{Alice Chess}{Popular chess variant invented by V. R. Parton.
See \url{https://en.wikipedia.org/wiki/Alice_chess}.}
\item{Alien City}{Awarding winning game by Michael Schoessow that is traditionally played with both a piecepack
and icehouse pieces.  See \url{https://www.ludism.org/ppwiki/AlienCity}.}
\item{Alquerque}{Traditional game also known as Qirkat.  See \url{https://en.wikipedia.org/wiki/Alquerque}.}
\item{Awithlaknannai Mosona}{Traditional Zuni Alquerque variant.
See \url{https://en.wikipedia.org/wiki/Awithlaknannai_Mosona}.}
\item{Backgammon}{One of the oldest board games.  See \url{https://www.ludism.org/ppwiki/Backgammon}.}
\item{Black Pawn Trucking}{A solitaire by Philip Lerche.
See \url{https://www.ludism.org/ppwiki/BlackPawnTrucking}.}
\item{Brain Burn}{A solitaire by Mark A. Biggar.
See \url{https://www.ludism.org/ppwiki/BrainBurn}.}
\item{Brandubh}{AKA Brandub, a two-player abstract of the hnefatafl family played by the Irish.
See \url{http://www.cyningstan.com/game/125/brandub} for more information.}
\item{Breakthrough}{Two-player abstract invented by Dan Troyka in 2000
that won the 2001 8x8 Game Design Competition.
See \url{https://en.wikipedia.org/wiki/Breakthrough_(board_game)}.}
\item{Burbuja}{Solitaire by Antonio Recuenco Muñoz.
See \url{https://www.ludism.org/ppwiki/Burbuja}.}
\item{Cardinal's Guards}{A solitaire by Michael Schoessow and Stephen Schoessow.
See \url{https://ludism.org/ppwiki/CardinalsGuards}.}
\item{Cell Management}{Solitaire game for the piecepack by Mark Goadrich.
See \url{https://www.ludism.org/ppwiki/CellManagement}.}
\item{Change Change}{Solitaire by Sid Sackson adapted by Ron Hale-Evans.
See \url{https://www.ludism.org/ppwiki/ChangeChange}.}
\item{Chariots}{Multiplayer racing game by Mark A. Biggar.
See \url{https://www.ludism.org/ppwiki/Chariots}.}
\item{Chaturaji}{An old 4-player chess variant.
See \url{https://www.ludism.org/ppwiki/Chaturaji}.}
\item{(American) Checkers}{A traditional board game also known as \dQuote{(English) Draughts}.
First adapted to the piecepack by Mark A. Biggar.
See \url{https://www.ludism.org/ppwiki/Checkers}.}
\item{(International) Chess}{Very popular board game first adapted to the piecepack by Ron Hale-Evans.
See \url{https://www.ludism.org/ppwiki/Chess}.}
\item{Piecepack Halma aka Chinese Checkers}{A port of Halma by Mark A. Biggar.
See \url{https://www.ludism.org/ppwiki/ChineseCheckers}.}
\item{Coin Collectors}{A solitaire by Don Kirkby.
See \url{https://www.ludism.org/ppwiki/CoinCollectors}.}
\item{Cribbage}{Traditional card game traditionally uses a special board to keep score
but one can use a piecepack as a cribbage board instead.
See \url{https://www.ludism.org/ppwiki/Cribbage}.}
\item{Climbing Man}{A solitaire by Jorge Arroyo.  See \url{https://www.ludism.org/ppwiki/ClimbingMan}.}
\item{Crocodile Hop}{A solitaire by Tim Schutz.  See \url{https://www.ludism.org/ppwiki/CrocodileHop}.}
\item{Crossings}{An abstract invented by Robert Abbot.
See \url{https://en.wikipedia.org/wiki/Crossings_(game)}}
\item{Desfases}{Game by Antonio Recuenco Muñoz.
Runner-up for best game in the eighth community piecepack game design contest.
See \url{https://www.ludism.org/ppwiki/Desfases}.}
\item{Easy Slider}{A solitaire by Ron Hale-Evans and Marty Hale-Evans.
See \url{https://www.ludism.org/ppwiki/EasySlider}.}
\item{Evade}{Adaption by Mark A. Biggar of a game by Alex Randolph.
See \url{https://www.ludism.org/ppwiki/Evade}.}
\item{Everest}{Game by Mark A. Biggar that features a Roborally programmed movement system.
See \url{https://www.ludism.org/ppwiki/Everest}.}
\item{Four Blind Mice}{Multi-player puzzle by Tim Farley inspired by Ricochet Robot by Alex Randolph.
See \url{https://www.ludism.org/ppwiki/FourBlindMice}.}
\item{Four Field Kono}{Traditional 2-player Korean abstract adapted to piecepack by Michael Schoessow.
See \url{https://www.ludism.org/ppwiki/FourFieldKono}.}
\item{Four Seasons Chess}{A traditional four person chess variant recorded in \emph{Libro del Acedrex}.
See \url{https://www.chessvariants.com/historic.dir/4seiz.html}.}
\item{Froggy Bottom}{Game by Clay Blankenship.
See \url{https://www.ludism.org/ppwiki/FroggyBottom}.}
\item{Fujisan}{Popular solitaire game for the piecepack by James \dQuote{Kyle} Droscha.
See \url{https://www.ludism.org/ppwiki/Fuji-san}.}
\item{Galaxy Express}{Solitaire by Matt Worden about delivering shipments between planets in the distant future.
See \url{https://www.ludism.org/ppwiki/GalaxyExpress}.}
\item{Grasshopper}{Two-player Halma variant playable on a draughts board.
See \url{http://www.cyningstan.com/game/71/grasshopper}.}
\item{Iceberg}{Game by Ken Leyhe.
See \url{https://www.ludism.org/ppwiki/Iceberg}.}
\item{Ice Floe}{Game by Tim Schutz requiring a piecepack and piecepack pyramids.
See \url{https://www.ludism.org/ppwiki/IceFloe}.}
\item{Japan}{Game by Daniel Ajoy and María Fernanda Ausay.
See \url{https://www.ludism.org/ppwiki/Japan}.}
\item{Jul-Gonu}{Traditional 2-player Korean abstract adapted to piecepack by Michael Schoessow.
See \url{https://www.ludism.org/ppwiki/JulGonu}.}
\item{Lab Rats}{Piecepack solitaire by Mark A. Biggar.
See \url{https://www.ludism.org/ppwiki/LabRats}.}
\item{Landlocked}{Piecepack game by Eric Witt.
See \url{https://www.ludism.org/ppwiki/Landlocked}.}
\item{Ley Lines}{Piecepack game by James \dQuote{Kyle} Droscha.
See \url{https://www.ludism.org/ppwiki/LeyLines}.}
\item{Lines of Action}{An abstract designed by Claude Soucie.
See \url{https://en.wikipedia.org/wiki/Lines_of_Action}.}
\item{Ludo}{Racing game based off Pachisi.
See \url{https://en.wikipedia.org/wiki/Ludo_(board_game)}.}
\item{Mathrix}{A solitaire by Clark Rodeffer.
See \url{https://www.ludism.org/ppwiki/Mathrix}}
\item{Minishogi}{A smaller version of Japanese chess invented (or perhaps rediscovered) by
by Shigeo Kusumoto in 1970.}
\item{Nine Men's Morris}{Traditional board game.
See \url{https://en.wikipedia.org/wiki/Nine_men\%27s_morris}.}
\item{One Man Thrag}{A solitaire by Jim Doherty.
See \url{https://www.ludism.org/ppwiki/OneManThrag}}
\item{Pass the Food}{A dexterity game by Trevor L Davis.
See \url{https://www.ludism.org/ppwiki/PassTheFood}}
\item{Piece Gaps}{A solitaire by Chris Brooks inspired by the Gaps card game.
See \url{https://www.ludism.org/ppwiki/PieceGaps}}
\item{Piece Packing Pirates}{A solitaire sea adventure by Clark Rodeffer.
See \url{https://www.ludism.org/ppwiki/PiecePackingPirates}.}
\item{Piecepack Klondike}{A solitaire game by Eric Witt inspired by the Klondike card game.
See \url{https://ludism.org/ppwiki/PiecepackKlondike}}
\item{Piecepackman}{A cooperative maze game by Dan Burkey inspired by the video game
Pac-Man, designed for Namco by Toru Iwatani.
See \url{https://www.ludism.org/ppwiki/Piecepackman}}
\item{Plans Of Action}{Solitaire piecepack game by L. Edward Pulley.
See \url{https://www.ludism.org/ppwiki/PlansOfAction}.}
\item{Quatri}{An abstract adapted to piecepack by Michael Schoessow.
See \url{https://www.ludism.org/ppwiki/Quatri}.}
\item{Relativity}{Piecepack game by Marty and Ron Hale-Evans.
See \url{https://www.ludism.org/ppwiki/Relativity}.}
\item{Salta}{Two-player abstract invented in 1899 by Konrad Heinrich Büttgenbach.
See \url{https://en.wikipedia.org/wiki/Salta_(game)}.}
\item{San Andreas}{Piecepack game about the big earthquake that will hit California.
See \url{https://www.ludism.org/ppwiki/SanAndreas}.}
\item{Sarcophagus}{A multiplayer piecepack game by Phillip Lerche.
See \url{https://www.ludism.org/ppwiki/Sarcophagus}.}
\item{Shogi}{AKA \dQuote{Japanese chess} is a major chess variant.
See \url{https://www.ludism.org/ppwiki/Shogi}.}
\item{Shopping Mall}{A multiplayer game by Michael Schoessow and Stephen Schoessow.
See \url{https://www.ludism.org/ppwiki/ShoppingMall}.}
\item{Skyscrapers}{A solitaire by Michael Schoessow.
See \url{https://www.ludism.org/ppwiki/Skyscrapers}.}
\item{Slides of Action}{An abstract connection game by Clark Rodeffer.
See \url{https://www.ludism.org/ppwiki/SlidesOfAction}.}
\item{Speedy Towers}{A real-time tower building game by Jessica Eccles.
See \url{https://ludism.org/ppwiki/SpeedyTowers}.}
\item{Steppin' Stones}{A solitaire by Ken Maher.
See \url{https://ludism.org/ppwiki/Steppin'_Stones}.}
\item{Tablut}{Traditional two-player abstract played by the Sámi people until at least the 1700s.
See \url{https://www.ludism.org/ppwiki/Tablut}.}
\item{The \dQuote{In} Crowd}{Piecepack game by Jeb Havens and Ian Schreiber.
See \url{https://www.ludism.org/ppwiki/TheInCrowd}.}
\item{The Magic Bag}{Piecepack game by Rob LeGood.
See \url{https://www.ludism.org/ppwiki/The_Magic_Bag}.}
\item{The Penguin Game}{A piecepack game by Jonathan C. Dietrich.
See \url{https://www.ludism.org/ppwiki/PenguinGame}}
\item{Tower of Babel}{Solitaire piecepack game by Mark A. Biggar.
See \url{https://www.ludism.org/ppwiki/TowerOfBabel}.}
\item{Tracers}{A two-player abstract by Marty Hale-Evans.
See \url{https://www.ludism.org/ppwiki/Tracers}.}
\item{Triactor}{Piecepack game by Jonathan Dietrich, Julie Taylor, and Ken MacKeigan.
See \url{https://www.ludism.org/ppwiki/Triactor}.}
\item{Tula}{Solitaire piecepack game by James \dQuote{Kyle} Droscha.
See \url{https://www.ludism.org/ppwiki/Tula}.}
\item{Turkish Draughts}{AKA \dQuote{Dama} is a traditional checkers variant played in the Mediterranean.
See \url{https://en.wikipedia.org/wiki/Turkish_draughts}.}
\item{Twelve Men's Morris}{Traditional two-player abstract.
See \url{https://en.wikipedia.org/wiki/Morabaraba}.}
\item{Ultima}{AKA \dQuote{Baroque chess}, a chess variant by Robert Abbott.
See \url{https://en.wikipedia.org/wiki/Baroque_chess}.}
\item{Wormholes}{Piecepack game by Marty and Ron Hale-Evans.
See \url{https://www.ludism.org/ppwiki/Wormholes}.}
\item{Xiangqi}{AKA \dQuote{Chinese chess}, a major chess variant.
See \url{https://www.ludism.org/ppwiki/Xiangqi}.}
}
}
