% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_kit.R
\name{game_kit}
\alias{game_kit}
\title{Game Kit R6 object}
\usage{
game_kit(cfgs = list(cfg = pp_cfg()))
}
\arguments{
\item{cfgs}{A named list of \code{pp_cfg} configuration list R6 objects.}
}
\description{
\code{game_kit} creates a game kit R6 object.
}
