% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_game.R
\name{plot_move}
\alias{plot_move}
\title{Plot game move}
\usage{
plot_move(
  game,
  file = NULL,
  move = NULL,
  annotate = TRUE,
  ...,
  .f = piecepackr::grid.piece,
  cfg = NULL,
  envir = NULL,
  width = NULL,
  height = NULL,
  ppi = 72,
  bg = "white",
  new_device = TRUE
)
}
\arguments{
\item{game}{A list containing a parsed ppn game (as parsed by \code{read_ppn})}

\item{file}{Filename to save animation unless \code{NULL}
in which case it uses the current graphics device.}

\item{move}{Which move to plot game state (after the move, will use \code{game$dfs[[move]]})
unless \code{NULL} in which case will plot the game state after the last move.}

\item{annotate}{If \code{TRUE} annotate the plot with \dQuote{algrebraic} coordinates,
if \code{FALSE} don't annotate,
if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.}

\item{...}{Arguments to \code{pmap_piece}}

\item{.f}{Low level graphics function to use e.g. \code{grid.piece}, \code{piece3d}, or \code{piece}.}

\item{cfg}{A piecepackr configuration list}

\item{envir}{Environment (or named list) of piecepackr configuration lists}

\item{width}{Width of animation (in inches).  Inferred by default.}

\item{height}{Height of animation (in inches).  Inferred by default.}

\item{ppi}{Resolution of animation in pixels per inch.
By default set so image max 600 pixels wide or tall.}

\item{bg}{Background color (\code{"transparent")} for transparent}

\item{new_device}{If \code{file} is \code{NULL} should we open up a new graphics device?}
}
\value{
Nothing, as a side effect saves a graphic
}
\description{
Plot game move
}
