% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppgames-defunct.R
\name{ppgames-defunct}
\alias{ppgames-defunct}
\alias{grid.board_rect_cells}
\alias{grid.board_rect_points}
\alias{grid.board_rect_tiles}
\title{Defunct functions}
\usage{
grid.board_rect_cells(...)

grid.board_rect_points(...)

grid.board_rect_tiles(...)
}
\arguments{
\item{...}{Not used}
}
\description{
These functions are Defunct in this release of ppgames,
they have been removed.
}
\details{
\enumerate{
\item{For \code{grid.board_rect_cells)}}
\item{For \code{grid.board_rect_points}}
\item{For \code{grid.board_rect_tiles}}
}
}
