% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{save_rulebook}
\alias{save_rulebook}
\title{Save rulebook}
\usage{
save_rulebook(
  book = "The Historical Piecepacker",
  gk = game_kit(),
  output = NULL,
  quietly = TRUE,
  size = "letter",
  cmd_options = NULL
)
}
\arguments{
\item{book}{Book name.  Currently only supports "The Historical Piecepacker".}

\item{gk}{A \code{game_kit} R6 object.}

\item{output}{Path to the output file.
If \code{NULL} the function will guess a default.}

\item{quietly}{Whether to hide document compilation output.}

\item{size}{Paper size (either "letter", or "A4").}

\item{cmd_options}{Options passed to \code{pandoc} when using non-pdf output formats.
If \code{NULL} we try to guess a good set of options.}
}
\description{
\code{save_rulebook} saves a rulebook.
}
