% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{save_ruleset}
\alias{save_ruleset}
\alias{save_rulebook}
\title{Save ruleset and/or rulebook}
\usage{
save_ruleset(game, gk = game_kit(), output_dir = getwd(), quietly = TRUE)

save_rulebook(
  book = "the-historical-piecepacker",
  gk = game_kit(),
  output_dir = getwd(),
  quietly = TRUE
)
}
\arguments{
\item{game}{Game name}

\item{gk}{A \code{game_kit} R6 object.}

\item{output_dir}{Directory to save ruleset/rulebook to.}

\item{quietly}{Whether to hide xelatex output.}

\item{book}{Book name}
}
\description{
Save ruleset and/or rulebook
}
