% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_game.R
\name{view_game}
\alias{view_game}
\title{View/edit game}
\usage{
view_game(
  game,
  shiny = FALSE,
  ...,
  editor = getOption("editor"),
  reorient = "none",
  annotate = FALSE,
  fps = 1
)
}
\arguments{
\item{game}{A list containing a parsed ppn game (as parsed by \code{read_ppn})}

\item{shiny}{If \code{TRUE} launch a shiny PPN viewer in a browser instead of command-line viewer.}

\item{...}{Passed to \code{plot_move()}.}

\item{editor}{usually a character string naming (or giving the path to) the text editor you want to use.}

\item{reorient}{Determines whether and how we should reorient (the angle) of pieces or symbols:\enumerate{
\item{The default "none" (or \code{FALSE}) means don't reorient any pieces/symbols.}
\item{"all" (or \code{TRUE}) means setting the angle to zero for all pieces
(reorienting them all \dQuote{up}).}
\item{"symbols" means just re-orient suit/rank symbols but not the orientation of the piece itself.
In particular, in contrast with "all" this preserves the location
of the upper-left "corner" of piecepack tile faces.}}}

\item{annotate}{If \code{TRUE} or \code{"algebraic"} annotate the plot
with \dQuote{algrebraic} coordinates,
if \code{FALSE} or \code{"none"} don't annotate,
if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.}

\item{fps}{Frames per second.  Passed to \code{cat_game()}.}
}
\description{
Launch PPN game viewer/editor
}
