% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_game.R
\name{animate_game}
\alias{animate_game}
\title{Animate a ppn game}
\usage{
animate_game(
  game,
  file = "animation.gif",
  annotate = TRUE,
  ...,
  cfg = NULL,
  envir = NULL,
  n_transitions = 0L,
  n_pauses = 1L,
  fps = n_transitions + n_pauses
)
}
\arguments{
\item{game}{A list containing a parsed ppn game (as parsed by \code{read_ppn})}

\item{file}{Filename to save animation}

\item{annotate}{If \code{TRUE} annotate the plot}

\item{...}{Arguments to \code{pmap_piece}}

\item{cfg}{A piecepackr configuration list}

\item{envir}{Environment (or named list) of piecepackr configuration lists}

\item{n_transitions}{Integer, if over zero (the default)
how many transition frames to add between moves.}

\item{n_pauses}{Integer, how many paused frames per completed move.}

\item{fps}{Double, frames per second.}
}
\value{
Nothing, as a side effect saves an animation of ppn game
}
\description{
Animate a ppn game
}
