% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_game.R
\name{animate_game}
\alias{animate_game}
\title{Animate a ppn game}
\usage{
animate_game(
  game,
  file = "animation.gif",
  annotate = TRUE,
  ...,
  .f = piecepackr::grid.piece,
  cfg = NULL,
  envir = NULL,
  n_transitions = 0L,
  n_pauses = 1L,
  fps = n_transitions + n_pauses,
  width = NULL,
  height = NULL,
  ppi = NULL,
  new_device = TRUE
)
}
\arguments{
\item{game}{A list containing a parsed ppn game (as parsed by \code{read_ppn})}

\item{file}{Filename to save animation unless \code{NULL}
in which case it uses the current graphics device.}

\item{annotate}{If \code{TRUE} or \code{"algebraic"} annotate the plot
with \dQuote{algrebraic} coordinates,
if \code{FALSE} or \code{"none"} don't annotate,
if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.}

\item{...}{Arguments to \code{pmap_piece}}

\item{.f}{Low level graphics function to use e.g. \code{grid.piece}, \code{piece3d}, or \code{piece}.}

\item{cfg}{A piecepackr configuration list}

\item{envir}{Environment (or named list) of piecepackr configuration lists}

\item{n_transitions}{Integer, if over zero (the default)
how many transition frames to add between moves.}

\item{n_pauses}{Integer, how many paused frames per completed move.}

\item{fps}{Double, frames per second.}

\item{width}{Width of animation (in inches).  Inferred by default.}

\item{height}{Height of animation (in inches).  Inferred by default.}

\item{ppi}{Resolution of animation in pixels per inch.
By default set so image max 600 pixels wide or tall.}

\item{new_device}{If \code{file} is \code{NULL} should we open up a new graphics device?}
}
\value{
Nothing, as a side effect saves an animation of ppn game
}
\description{
Animate a ppn game
}
