% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{boards}
\alias{boards}
\title{Various game boards}
\arguments{
\item{nrows}{Number of rows in game board}

\item{ncols}{Number of columns in game board}

\item{x0}{X coordinate for the center of the first cell/point}

\item{y0}{Y coordinate for the center of the first cell/point}

\item{max_tiles}{Maximum number of tiles that can be used}

\item{cfg}{Piecepack configuration list or \code{pp_cfg} object, 
a list of \code{pp_cfg} objects,
or a character vector of \code{pp_cfg} objects}

\item{envir}{Environment (or named list) containing configuration list(s).}

\item{name}{A character identifier (for grid)}

\item{gp}{An object of class ‘gpar’, typically the output from a call
to the function ‘gpar’.  This is basically a list of
graphical parameter settings.}

\item{draw}{A logical value indicating whether graphics output should be produced.}

\item{vp}{A \code{grid} viewport object (or NULL).}
}
\description{
\code{grid.board_rect_tiles} draws a rectangular board using the back of piecepack tiles, 
it applies \code{pmap_piece} to the \code{tibble} data frame returned by \code{df_rect_board_tiles}.
\code{grid.board_rect_cells} draws a rectangular board where pieces are placed in the cells of the board.
\code{grid.board_rect_points} draws a rectangular board where pieces are placed in the points of the board.
}
