% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicode.R
\name{cat_piece}
\alias{cat_piece}
\title{Generate plaintext piecepack diagrams}
\usage{
cat_piece(df, ...)
}
\arguments{
\item{df}{Data frame containing piece info.}

\item{...}{Passed to \code{cat}}
}
\value{
None (invisible \code{NULL})
}
\description{
\code{format_piece} generates plaintext piecepack diagrams and 
 outputs them using \code{cat}.
}
