% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicode.R
\name{cat_piece}
\alias{cat_piece}
\alias{cat_move}
\alias{cat_game}
\title{Generate plaintext piecepack diagrams}
\usage{
cat_piece(
  df,
  color = NULL,
  reorient = "none",
  annotate = FALSE,
  ...,
  file = "",
  annotation_scale = NULL
)

cat_move(game, move = NULL, ...)

cat_game(game, ..., fps = 1)
}
\arguments{
\item{df}{Data frame containing piece info.}

\item{color}{How should the text be colorized.
If \code{FALSE} won't colorize output at all.
If \code{"html"} will colorize output for html.
Otherwise will colorize output for the terminal using ANSI CSI SGR control sequences.}

\item{reorient}{Determines whether and how we should reorient (the angle) of pieces or symbols:\enumerate{
\item{The default "none" (or \code{FALSE}) means don't reorient any pieces/symbols.}
\item{"all" (or \code{TRUE}) means setting the angle to zero for all pieces
(reorienting them all \dQuote{up}).}
\item{"symbols" means just re-orient suit/rank symbols but not the orientation of the piece itself.
In particular, in contrast with "all" this preserves the location
of the upper-left "corner" of piecepack tile faces.}}}

\item{annotate}{If \code{TRUE} or \code{"algebraic"} annotate the plot
with \dQuote{algrebraic} coordinates,
if \code{FALSE} or \code{"none"} don't annotate,
if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.}

\item{...}{Passed to \code{cat()}}

\item{file}{\code{file} argument of \code{cat()}.
Default (\code{""}) is to print to standard output.
\code{NULL} means we don't \code{cat()}}

\item{annotation_scale}{Multiplicative factor that scales (stretches) any annotation coordinates.
By default uses \code{attr(df, "scale_factor") \%||\% 1}.}

\item{game}{A list containing a parsed ppn game (as parsed by \code{read_ppn})}

\item{move}{Which move to cat game state (after the move, will use \code{game$dfs[[move]]})
unless \code{NULL} in which case will cat the game state after the last move.}

\item{fps}{Frames per second.}
}
\value{
String of text diagram (returned invisibly).
}
\description{
\code{cat_piece()} generates plaintext piecepack diagrams and
outputs them using \code{base::cat()}.  \code{cat_move()} generates
a plaintext diagram for a move within a game.  \code{cat_game()}
renders an animation of a game in the terminal.
}
