% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R, R/game_starts_chess.R,
%   R/game_starts_traditional.R
\name{df_game}
\alias{df_game}
\alias{df_alien_city}
\alias{df_cell_management}
\alias{df_desfases}
\alias{df_everest}
\alias{df_fujisan}
\alias{df_ice_floe}
\alias{df_plans_of_action}
\alias{df_relativity}
\alias{df_the_in_crowd}
\alias{df_san_andreas}
\alias{df_triactor}
\alias{df_wormholes}
\alias{df_alice_chess}
\alias{df_chaturaji}
\alias{df_four_seasons_chess}
\alias{df_international_chess}
\alias{df_chess}
\alias{df_shogi}
\alias{df_ultima}
\alias{df_baroque_chess}
\alias{df_xiangqi}
\alias{df_american_checkers}
\alias{df_english_draughts}
\alias{df_checkers}
\alias{df_backgammon}
\alias{df_cribbage_board}
\alias{textGrob_cribbage_board}
\alias{df_four_field_kono}
\alias{df_nine_mens_morris}
\alias{df_twelve_mens_morris}
\alias{df_salta}
\alias{df_tablut}
\title{Data frames of starting diagrams for various games}
\usage{
df_alien_city(seed = NULL, tiles = NULL)

df_cell_management(seed = NULL)

df_desfases(seed = NULL, tiles = NULL, dice = NULL)

df_everest()

df_fujisan(seed = NULL, coins = NULL, dice = NULL)

df_ice_floe()

df_plans_of_action(seed = NULL, coins = NULL)

df_relativity(seed = NULL, coins = NULL)

df_the_in_crowd()

df_san_andreas()

df_triactor(seed = NULL, cfg2 = "playing_cards_expansion")

df_wormholes()

df_alice_chess(has_subpack = FALSE, max_tiles = 24)

df_chaturaji(has_subpack = FALSE)

df_four_seasons_chess(has_subpack = FALSE)

df_international_chess(has_subpack = FALSE)

df_chess(has_subpack = FALSE)

df_shogi(has_subpack = FALSE, cfg2 = "piecepack")

df_ultima(has_subpack = FALSE)

df_baroque_chess(has_subpack = FALSE)

df_xiangqi(has_subpack = FALSE)

df_american_checkers()

df_english_draughts()

df_checkers()

df_backgammon()

df_cribbage_board()

textGrob_cribbage_board(suit_colors = rep("black", 4))

df_four_field_kono()

df_nine_mens_morris(has_matchsticks = FALSE)

df_twelve_mens_morris(has_matchsticks = FALSE)

df_salta(has_subpack = FALSE)

df_tablut(die_width = 0.5)
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{tiles}{String of tile layout}

\item{dice}{String of dice layout}

\item{coins}{String of coin layout}

\item{cfg2}{A string of a piecepack expansion (or perhaps \code{"piecepack"} for a second piecepack)}

\item{has_subpack}{Has a piecepack subpack}

\item{max_tiles}{Maximum number of (piecepack) tiles available to build boards}

\item{suit_colors}{Character vector of the suit colors}

\item{has_matchsticks}{Has matchsticks}

\item{die_width}{Width of dice}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
  data frame output can usually be plotted with \code{pmap_piece(df, cfg = cfg, default.units = "in")}.
}
