% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R
\name{df_game}
\alias{df_game}
\alias{df_backgammon}
\alias{df_tablut}
\alias{df_cribbage_board}
\title{Data frames of starting diagrams for various games}
\usage{
df_backgammon()

df_tablut(cfg = pp_cfg())

df_cribbage_board()
}
\arguments{
\item{cfg}{A \code{pp_cfg} configuration list object}

\item{cfg2}{A \code{pp_cfg} configuration list object}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
  data frame output can usually be plotted with \code{pmap_piece(df, cfg=cfg, default.units="in")}.
}
