% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R
\name{df_game}
\alias{df_game}
\alias{df_alien_city}
\alias{df_everest}
\alias{df_four_field_kono}
\alias{df_cell_management}
\alias{df_fujisan}
\alias{df_nine_mens_morris}
\alias{df_twelve_mens_morris}
\alias{df_american_checkers}
\alias{df_backgammon}
\alias{df_chaturaji}
\alias{df_cribbage_board}
\alias{textGrob_cribbage_board}
\alias{df_fide_chess}
\alias{df_ultima}
\alias{df_baroque_chess}
\alias{df_alice_chess}
\alias{df_four_seasons_chess}
\alias{df_shogi}
\alias{df_tablut}
\alias{df_triactor}
\alias{df_xiangqi}
\title{Data frames of starting diagrams for various games}
\usage{
df_alien_city(seed = NULL, tiles = NULL)

df_everest()

df_four_field_kono()

df_cell_management(seed = NULL)

df_fujisan(seed = NULL, coins = NULL, dice = NULL)

df_nine_mens_morris(has_matchsticks = FALSE)

df_twelve_mens_morris(has_matchsticks = FALSE)

df_american_checkers()

df_backgammon()

df_chaturaji(has_subpack = FALSE)

df_cribbage_board()

textGrob_cribbage_board(suit_colors = rep("black", 4))

df_fide_chess(has_subpack = FALSE)

df_ultima(has_subpack = FALSE)

df_baroque_chess(has_subpack = FALSE)

df_alice_chess(has_subpack = FALSE, max_tiles = 24)

df_four_seasons_chess(has_subpack = FALSE)

df_shogi(has_subpack = FALSE, cfg2 = "piecepack")

df_tablut(die_width = 0.5)

df_triactor(seed = NULL, cfg2 = "playing_cards_expansion")

df_xiangqi(has_subpack = FALSE)
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{tiles}{String of tile layout}

\item{coins}{String of coin layout}

\item{dice}{String of dice layout}

\item{has_matchsticks}{Has matchsticks}

\item{has_subpack}{Has a piecepack subpack}

\item{suit_colors}{Character vector of the suit colors}

\item{max_tiles}{Maximum number of (piecepack) tiles available to build boards}

\item{cfg2}{A string of a piecepack expansion (or perhaps \code{"piecepack"} for a second piecepack)}

\item{die_width}{Width of dice}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
  data frame output can usually be plotted with \code{pmap_piece(df, cfg=cfg, default.units="in")}.
}
