% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_starts.R, R/game_starts_chess.R,
%   R/game_starts_traditional.R
\name{df_game}
\alias{df_game}
\alias{df_alien_city}
\alias{df_cell_management}
\alias{df_desfases}
\alias{df_everest}
\alias{df_froggy_bottom}
\alias{df_fujisan}
\alias{df_ice_floe}
\alias{df_ley_lines}
\alias{df_piecepackman}
\alias{df_plans_of_action}
\alias{df_relativity}
\alias{df_the_in_crowd}
\alias{df_san_andreas}
\alias{df_triactor}
\alias{df_wormholes}
\alias{df_alice_chess}
\alias{df_chaturaji}
\alias{df_four_seasons_chess}
\alias{df_international_chess}
\alias{df_chess}
\alias{df_shogi}
\alias{df_ultima}
\alias{df_baroque_chess}
\alias{df_xiangqi}
\alias{df_american_checkers}
\alias{df_english_draughts}
\alias{df_checkers}
\alias{df_backgammon}
\alias{df_breakthrough}
\alias{df_cribbage_board}
\alias{textGrob_cribbage_board}
\alias{df_crossings}
\alias{df_four_field_kono}
\alias{df_lines_of_action}
\alias{df_nine_mens_morris}
\alias{df_turkish_draughts}
\alias{df_twelve_mens_morris}
\alias{df_salta}
\alias{df_tablut}
\title{Data frames of starting diagrams for various games}
\usage{
df_alien_city(seed = NULL, tiles = NULL)

df_cell_management(seed = NULL)

df_desfases(seed = NULL, tiles = NULL, dice = NULL)

df_everest()

df_froggy_bottom()

df_fujisan(seed = NULL, coins = NULL, dice = NULL)

df_ice_floe()

df_ley_lines()

df_piecepackman(seed = NULL, variant = 1)

df_plans_of_action(seed = NULL, coins = NULL)

df_relativity(seed = NULL, coins = NULL)

df_the_in_crowd()

df_san_andreas()

df_triactor(seed = NULL, cfg2 = "playing_cards_expansion")

df_wormholes()

df_alice_chess(has_subpack = FALSE, max_tiles = 24)

df_chaturaji(has_subpack = FALSE)

df_four_seasons_chess(has_subpack = FALSE)

df_international_chess(has_subpack = FALSE)

df_chess(has_subpack = FALSE)

df_shogi(has_subpack = FALSE, cfg2 = "piecepack")

df_ultima(has_subpack = FALSE)

df_baroque_chess(has_subpack = FALSE)

df_xiangqi(has_subpack = FALSE)

df_american_checkers()

df_english_draughts()

df_checkers()

df_backgammon()

df_breakthrough()

df_cribbage_board()

textGrob_cribbage_board(suit_colors = rep("black", 4))

df_crossings()

df_four_field_kono()

df_lines_of_action()

df_nine_mens_morris(has_matchsticks = FALSE)

df_turkish_draughts()

df_twelve_mens_morris(has_matchsticks = FALSE)

df_salta(has_subpack = FALSE)

df_tablut(die_width = 0.63)
}
\arguments{
\item{seed}{Seed that determines setup, either an integer or \code{NULL}}

\item{tiles}{String of tile layout}

\item{dice}{String of dice layout}

\item{coins}{String of coin layout}

\item{variant}{Number of variant.}

\item{cfg2}{A string of a piecepack expansion (or perhaps \code{"piecepack"} for a second piecepack)}

\item{has_subpack}{Has a piecepack subpack}

\item{max_tiles}{Maximum number of (piecepack) tiles available to build boards}

\item{suit_colors}{Character vector of the suit colors}

\item{has_matchsticks}{Has matchsticks}

\item{die_width}{Width of dice}
}
\description{
\code{tibble} data frames of starting diagrams for various games.
  data frame output can usually be plotted with \code{pmap_piece(df, cfg = cfg, default.units = "in")}.
}
\details{
Here is some more information about the various games:
  \describe{
  \item{Alice Chess}{Popular chess variant invented by V. R. Parton.
                     See \url{https://en.wikipedia.org/wiki/Alice_chess}.}
 \item{Alien City}{Awarding winning game by Michael Schoessow that is traditionally played with both a piecepack
                   and icehouse pieces.  See \url{http://www.ludism.org/ppwiki/AlienCity}.}
 \item{Backgammon}{One of the oldest board games.  See \url{http://www.ludism.org/ppwiki/Backgammon}.}
 \item{Breakthrough}{Two-player abstract invented by Dan Troyka in 2000
                     that won the 2001 8x8 Game Design Competition.
                     See \url{https://en.wikipedia.org/wiki/Breakthrough_(board_game)}.}
 \item{Cell Management}{Solitaire game for the piecepack by Mark Goadrich.
                        See \url{http://www.ludism.org/ppwiki/CellManagement}.}
 \item{Chaturaji}{An old 4-player chess variant.
                  See \url{http://www.ludism.org/ppwiki/Chaturaji}.}
 \item{(American) Checkers}{A traditional board game also known as \dQuote{(English) Draughts}.
                 First adapted to the piecepack by Mark A. Biggar.
                 See \url{http://www.ludism.org/ppwiki/Checkers}.}
 \item{(International) Chess}{Very popular board game first adapted to the piecepack by Ron Hale-Evans.
                              See \url{http://www.ludism.org/ppwiki/Chess}.}
 \item{Cribbage}{Traditional card game traditionally uses a special board to keep score
                 but one can use a piecepack as a cribbage board instead.
                 See \url{http://www.ludism.org/ppwiki/Cribbage}.}
 \item{Crossings}{An abstract invented by Robert Abbot.
                  See \url{https://en.wikipedia.org/wiki/Crossings_(game)}}
 \item{Desfases}{Game by Antonio Recuenco Muñoz.
                 Runner-up for best game in the eighth community piecepack game design contest.
                 See \url{http://www.ludism.org/ppwiki/Desfases}.}
 \item{Everest}{Game by Mark A. Biggar that features a Roborally programmed movement system.
                See \url{http://www.ludism.org/ppwiki/Everest}.}
 \item{Four Field Kono}{Traditional 2-player Korean abstract adapted to piecepack by Michael Schoessow.
                        See \url{http://www.ludism.org/ppwiki/FourFieldKono}.}
 \item{Four Seasons Chess}{A traditional four person chess variant recorded in \emph{Libro del Acedrex}.
                           See \url{https://www.chessvariants.com/historic.dir/4seiz.html}.}
 \item{Froggy Bottom}{Game by Clay Blankenship.
                      See \url{http://www.ludism.org/ppwiki/FroggyBottom}.}
 \item{Fujisan}{Popular solitaire game for the piecepack by James \dQuote{Kyle} Droscha.
                See \url{http://www.ludism.org/ppwiki/Fuji-san}.}
 \item{Ice Floe}{Game by Tim Schutz requiring a piecepack and piecepack pyramids.
                 See \url{http://www.ludism.org/ppwiki/IceFloe}.}
 \item{Ley Lines}{Piecepack game by James \dQuote{Kyle} Droscha.
                  See \url{http://www.ludism.org/ppwiki/LeyLines}.}
 \item{Lines of Action}{An abstract designed by Claude Soucie.
                        See \url{https://en.wikipedia.org/wiki/Lines_of_Action}.}
 \item{Nine Men's Morris}{Traditional board game.
       See \url{https://en.wikipedia.org/wiki/Nine_men\%27s_morris}.}
 \item{Piecepackman}{A cooperative maze game by Dan Burkey inspired by the video game
                     Pac-Man, designed for Namco by Toru Iwatani}
 \item{Plans Of Action}{Solitaire piecepack game by L. Edward Pulley.
       See \url{http://www.ludism.org/ppwiki/PlansOfAction}.}
 \item{Relativity}{Piecepack game by Marty and Ron Hale-Evans.
                   See \url{http://www.ludism.org/ppwiki/Relativity}.}
 \item{Salta}{Two-player abstract invented in 1899 by Konrad Heinrich Büttgenbach.
              See \url{https://en.wikipedia.org/wiki/Salta_(game)}.}
 \item{San Andreas}{Piecepack game about the big earthquake that will hit California.
                   See \url{http://www.ludism.org/ppwiki/SanAndreas}.}
 \item{Shogi}{AKA \dQuote{Japanese chess} is a major chess variant.
              See \url{http://www.ludism.org/ppwiki/Shogi}.}
 \item{Tablut}{Traditional two-player abstract played by the Sámi people until at least the 1700s.
               See \url{http://www.ludism.org/ppwiki/Tablut}.}
 \item{The \dQuote{In} Crowd}{Piecepack game by Jeb Havens and Ian Schreiber.
       See \url{http://www.ludism.org/ppwiki/TheInCrowd}.}
 \item{Triactor}{Piecepack game by Jonathan Dietrich, Julie Taylor, and Ken MacKeigan.
                 See \url{http://www.ludism.org/ppwiki/Triactor}.}
 \item{Turkish Draughts}{AKA \dQuote{Dama} is a traditional checkers variant played in the Mediterranean.
                         See \url{https://en.wikipedia.org/wiki/Turkish_draughts}.}
 \item{Twelve Men's Morris}{Traditional two-player abstract.
                            See \url{https://en.wikipedia.org/wiki/Morabaraba}.}
 \item{Ultima}{AKA \dQuote{Baroque chess}, a chess variant by Robert Abbott.
               See \url{https://en.wikipedia.org/wiki/Baroque_chess}.}
 \item{Xiangqi}{AKA \dQuote{Chinese chess}, a major chess variant.
                See \url{http://www.ludism.org/ppwiki/Xiangqi}.}
 \item{Wormholes}{Piecepack game by Marty and Ron Hale-Evans.
                  See \url{http://www.ludism.org/ppwiki/Wormholes}.}
}
}
