% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{names_rulesets}
\alias{names_rulesets}
\title{Get names of piecepack games we can generate rulesets for.}
\usage{
names_rulesets(book = NULL)
}
\arguments{
\item{book}{Book name or \code{NULL} (for all supported rules).
Currently only supports "the historical piecepacker".}
}
\description{
\code{names_rulesets()} returns the names of piecepack games we can generate rulesets for.
}
\seealso{
\code{\link[=save_ruleset]{save_ruleset()}}, \code{\link[=save_pamphlet]{save_pamphlet()}}, and \code{\link[=save_pocketmod]{save_pocketmod()}}.
}
