% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_move.R
\name{plot_move}
\alias{plot_move}
\title{Plot game move}
\usage{
plot_move(
  game,
  file = NULL,
  move = NULL,
  annotate = TRUE,
  ...,
  .f = piecepackr::grid.piece,
  cfg = NULL,
  envir = NULL,
  width = NULL,
  height = NULL,
  ppi = 72,
  bg = "white",
  new_device = TRUE,
  annotation_scale = NULL
)
}
\arguments{
\item{game}{A list containing a parsed ppn game (as parsed by \code{\link[=read_ppn]{read_ppn()}})}

\item{file}{Filename to save animation unless \code{NULL}
in which case it uses the current graphics device.}

\item{move}{Which move to plot game state (after the move, will use \code{game$dfs[[move]]})
unless \code{NULL} in which case will plot the game state after the last move.}

\item{annotate}{If \code{TRUE} or \code{"algebraic"} annotate the plot
with \dQuote{algrebraic} coordinates,
if \code{FALSE} or \code{"none"} don't annotate,
if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.}

\item{...}{Passed to \code{\link[piecepackr:render_piece]{piecepackr::render_piece()}}}

\item{.f}{Low level graphics function to use e.g. \code{\link[piecepackr:grid.piece]{grid.piece()}}, \code{\link[piecepackr:piece3d]{piece3d()}}, \code{\link[piecepackr:piece]{piece()}}, or \code{\link[piecepackr:piece_mesh]{piece_mesh()}}.}

\item{cfg}{A piecepackr configuration list}

\item{envir}{Environment (or named list) of piecepackr configuration lists}

\item{width}{Width of animation (in inches).  Inferred by default.}

\item{height}{Height of animation (in inches).  Inferred by default.}

\item{ppi}{Resolution of animation in pixels per inch.
By default set so image max 600 pixels wide or tall.}

\item{bg}{Background color (\code{"transparent")} for transparent}

\item{new_device}{If \code{file} is \code{NULL} should we open up a new graphics device?}

\item{annotation_scale}{Multiplicative factor that scales (stretches) any annotation coordinates.
By default uses \code{attr(df, "scale_factor") \%||\% 1}.}
}
\value{
An invisible list of the dimensions of the image, as a side effect saves a graphic
}
\description{
Plot game move
}
