% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_game.R
\name{plot_move}
\alias{plot_move}
\title{Plot game move}
\usage{
plot_move(
  game,
  file = NULL,
  move = NULL,
  annotate = TRUE,
  ...,
  bg = "white",
  res = 72,
  cfg = NULL,
  envir = NULL,
  new_device = TRUE
)
}
\arguments{
\item{game}{A list containing a parsed ppn game (as parsed by \code{read_ppn})}

\item{file}{Filename to save graphic to unless \code{NULL}
in which case it uses the current graphics device.}

\item{move}{Which move to plot game state (after the move, will use \code{game$dfs[[move]]})
unless \code{NULL} in which case will plot the game state after the last move.}

\item{annotate}{If \code{TRUE} annotate the plot with \dQuote{algrebraic} coordinates,
if \code{FALSE} don't annotate,
if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.}

\item{...}{Arguments to \code{pmap_piece}}

\item{bg}{Background color (\code{"transparent")} for transparent}

\item{res}{For bitmap image formats the resolution}

\item{cfg}{A piecepackr configuration list}

\item{envir}{Environment (or named list) of piecepackr configuration lists}

\item{new_device}{If \code{file} is \code{NULL} should we open up a new graphics device?}
}
\value{
Nothing, as a side effect saves a graphic
}
\description{
Plot game move
}
