% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppn.R
\name{read_ppn}
\alias{read_ppn}
\title{Read ppn files}
\usage{
read_ppn(file, parse = TRUE)
}
\arguments{
\item{file}{Filename}

\item{parse}{Logical of whether to parse the moves in the ppn file}
}
\value{
A list, for each game in the file a list containing info about the game
}
\description{
Read in ppn files
}
