% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{save_ruleset}
\alias{save_ruleset}
\alias{save_pamphlet}
\alias{save_rulebook}
\title{Save ruleset and/or rulebook}
\usage{
save_ruleset(
  game,
  gk = game_kit(),
  output = NULL,
  quietly = TRUE,
  size = "letter"
)

save_pamphlet(
  game,
  gk = game_kit(),
  output = NULL,
  quietly = TRUE,
  size = "letter",
  duplex_edge = "short"
)

save_rulebook(
  book = "The Historical Piecepacker",
  gk = game_kit(),
  output = NULL,
  quietly = TRUE,
  size = "letter"
)
}
\arguments{
\item{game}{Game name}

\item{gk}{A \code{game_kit} R6 object.}

\item{output}{Path to the output file.
If \code{NULL} the function will guess a default.}

\item{quietly}{Whether to hide document compilation output.}

\item{size}{Paper size (either "letter", or "A4")}

\item{duplex_edge}{String specifying the desired duplex printing edge.
If "short" match the second page along its short edge (second page flipped up, easier to preview on computer)
and if "long" match along its long edge (second page flipped upside down, usual printer default).}

\item{book}{Book name}
}
\description{
\code{save_ruleset} save ruleset of a game,
\code{save_pamphlet} is a variant that saves the ruleset as a pamphlet.
\code{save_rulebook} saves a rulebook.
}
